change directory and apply minor refactoring for config class
